app/ray_serve/requirements.txt
torch>=2.2.0,<2.7.0
transformers>=4.40.0,<5.0.0
Pillow>=10.0.0
librosa>=0.10.0
soundfile>=0.12.0
numpy<2.0.0
